pyproject.toml
[project]
name = "teledictionary_bot"
version = "0.1.0"
description = "Telegram bot that allows you to get definitions of various words from Oxford and Urban Dictionaries"
authors = [
    {name = "Pavlo Pohorieltsev", email = "49622129+makisukurisu@users.noreply.github.com"},
]
dependencies = [
    "python-telegram-bot>=20.8",
    "pydantic>=2.6.2",
    "pydantic-settings>=2.2.1",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "ruff>=0.2.2",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
start_bot={call = "teledictionary_bot.main:build", help="Starts the bot. (Executes build method of main module)"}
echo_done={shell = "echo Done!"}
_perform_lint="ruff check . --fix"
lint={composite = ["_perform_lint", "echo_done"], help="Executes ruff check, and echoes 'Done!'."}
format="ruff format ."
full_ruff={composite = ["format", "lint"], help="Runs formatter and linter on the code"}

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
extend-select = [
  "UP",
  "E",
  "W",
  "I",
  "C90",
  "N",
  "ANN",
  "ASYNC",
  "BLE",
  "FBT",
  "B",
  "A",
  "C4",
  "DTZ",
  "T10",
  "FA",
  "ISC",
  "ICN",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "Q",
  "RET",
  "SLF",
  "SIM",
  "TID",
  "TCH",
  "ARG",
  "TD",
  "FIX",
  "ERA",
  "PL",
  "RUF",
]
ignore = [
  "ISC001",
  "N999",
  "ERA001",
  "RUF012",
  "ANN101",
  "ANN102",
  "FBT001",
  "FBT002",
]

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true